utf/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-model-utils==4.5.1
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.12.2
